requirements.txt
Django==5.0
django_allauth==0.58.2
Pillow==10.1.0
python-decouple==3.8
django-widget-tweaks==1.5.0
psycopg2-binary==2.9.9
whitenoise==6.6.0